Update: UserRepository를 class로 변경
- Repository 변경 예정이 없기 때문에 interface제거
- 최초 회원가입 시 UserBadge 1개 발행하도록 추가
