--- src/main/java/com/divide/user/UserJpaRepository.java
@@ -1,43 +1,0 @@
-package com.divide.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.NoResultException;
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-@RequiredArgsConstructor
-public class UserJpaRepository implements UserRepository{
-    private final EntityManager em;
-
-    @Override
-    public List<User> getUsers() {
-        return em.createQuery("select u from User u", User.class).getResultList();
-    }
-
-    @Override
-    public void signup(User user) {
-        em.persist(user);
-    }
-
-    @Override
-    public Optional<User> findByEmail(String email){
-        try {
-            return Optional.of(em.createQuery("select u from User u" +
-                            " where u.email = :email", User.class)
-                    .setParameter("email", email)
-                    .getSingleResult());
-        } catch (NoResultException e) {
-            return Optional.empty();
-        }
-    }
-
-    @Override
-    public User findById(Long id){
-        User user = em.find(User.class, id); //postId는 PK
-        return user;
-    }
-}
